Split kubernetes deployment tests into a dedicated test file
